refactor(MovieListPresenter): Repository 추상화 의존성 주입, 패키지명 변경
