+++ app/src/main/AndroidManifest.xml
@@ -14,7 +14,7 @@
         android:theme="@style/Theme.Movie"
         tools:targetApi="31" >
         <receiver
-            android:name=".view.seatselection.AlarmReceiver"
+            android:name=".view.AlarmReceiver"
             android:enabled="true"
             android:exported="true" >
         </receiver>
