+++ app/src/main/java/woowacourse/movie/data/movie/MovieMockRepository.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.data
+package woowacourse.movie.data.movie
 
 import woowacourse.movie.domain.movie.Minute
 import woowacourse.movie.domain.movie.Movie
