+++ app/src/main/java/woowacourse/movie/data/movie/MoviePosterRepository.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.data
+package woowacourse.movie.data.movie
 
 interface MoviePosterRepository {
     fun findPoster(title: String): Int
